feedstock/requirements.txt
s3fs
boto3
python-cmr==0.9.0
pangeo-forge-recipes==0.10.3
